# Optimized RenderTexture performance