--- a/src/SFML/Graphics/RenderTextureImplFBO.hpp
+++ b/src/SFML/Graphics/RenderTextureImplFBO.hpp
@@ -31,6 +31,7 @@
 #include <SFML/Graphics/RenderTextureImpl.hpp>
 #include <SFML/Window/Context.hpp>
 #include <SFML/Window/GlResource.hpp>
+#include <map>
 
 
 namespace sf
@@ -74,6 +75,12 @@ public:
     ////////////////////////////////////////////////////////////
     static unsigned int getMaximumAntialiasingLevel();
 
+    ////////////////////////////////////////////////////////////
+    /// \brief Unbind the currently bound FBO
+    ///
+    ////////////////////////////////////////////////////////////
+    static void unbind();
+
 private:
 
     ////////////////////////////////////////////////////////////
@@ -88,6 +95,14 @@ private:
     ///
     ////////////////////////////////////////////////////////////
     virtual bool create(unsigned int width, unsigned int height, unsigned int textureId, const ContextSettings& settings);
+
+    ////////////////////////////////////////////////////////////
+    /// \brief Create an FBO in the current context
+    ///
+    /// \return True if creation has been successful
+    ///
+    ////////////////////////////////////////////////////////////
+    bool createFrameBuffer();
 
     ////////////////////////////////////////////////////////////
     /// \brief Activate or deactivate the render texture for rendering
@@ -110,13 +125,16 @@ private:
     ////////////////////////////////////////////////////////////
     // Member data
     ////////////////////////////////////////////////////////////
-    Context*     m_context;                ///< Needs a separate OpenGL context for not messing up the other ones
-    unsigned int m_frameBuffer;            ///< OpenGL frame buffer object
-    unsigned int m_multisampleFrameBuffer; ///< Optional OpenGL frame buffer object with multisample attachments
-    unsigned int m_depthStencilBuffer;     ///< Optional depth/stencil buffer attached to the frame buffer
-    unsigned int m_colorBuffer;            ///< Optional multisample color buffer attached to the frame buffer
-    unsigned int m_width;                  ///< Width of the attachments
-    unsigned int m_height;                 ///< Height of the attachments
+    std::map<Uint64, unsigned int> m_frameBuffers;            ///< OpenGL frame buffer objects per context
+    std::map<Uint64, unsigned int> m_multisampleFrameBuffers; ///< Optional per-context OpenGL frame buffer objects with multisample attachments
+    unsigned int                   m_depthStencilBuffer;      ///< Optional depth/stencil buffer attached to the frame buffer
+    unsigned int                   m_colorBuffer;             ///< Optional multisample color buffer attached to the frame buffer
+    unsigned int                   m_width;                   ///< Width of the attachments
+    unsigned int                   m_height;                  ///< Height of the attachments
+    Context*                       m_context;                 ///< Backup OpenGL context, used when none already exist
+    unsigned int                   m_textureId;               ///< The ID of the texture to attach to the FBO
+    bool                           m_multisample;             ///< Whether we have to create a multisample frame buffer as well
+    bool                           m_stencil;                 ///< Whether we have stencil attachment
 };
 
 } // namespace priv