… src/SFML/Graphics/RenderTextureImplFBO.cpp