Optimized RenderTexture performance when using the FBO implementation by removing unnecessary context switches and flushing.

In this change, FBOs no longer live in their own context if they don't have to. This reduces or even eliminates context switching all together in single window applications, giving "near native OpenGL performance". Furthermore, the glFlush() in display() was also removed in favour of explicit rebinding in RenderTarget in order to reduce the time waiting for the driver queue to flush to the GPU queue. This also saves a large amount of time.

In applications that make heavy use of RenderTextures, this can mean a significant performance gain if they were previously CPU-bottlenecked. This change shouldn't have an effect on the actual GPU workload in any way.

Here is a synthetic "benchmark" I used to test the difference with and without this change:
```cpp
#include <SFML/Graphics.hpp>
#include <sstream>

int main()
{
    sf::RenderWindow window(sf::VideoMode(800, 600), "Test", sf::Style::Titlebar | sf::Style::Close);
    window.setVerticalSyncEnabled(false);

    sf::Font font;
    if (!font.loadFromFile("resources/sansation.ttf"))
        return EXIT_FAILURE;

    sf::Text text("Test", font, 20);

    sf::RenderTexture renderTexture[10];

    for (int i = 0; i < 10; ++i)
        renderTexture[i].create(800, 600);

    sf::Sprite sprite;

    sf::Clock clock;
    std::ostringstream sstr;

    while (window.isOpen())
    {
        sf::Event event;
        while (window.pollEvent(event))
        {
            if (event.type == sf::Event::Closed)
                window.close();

            if ((event.type == sf::Event::KeyPressed) && (event.key.code == sf::Keyboard::Escape))
                window.close();
        }

        for (int i = 0; i < 20; ++i)
        {
            for (int j = 0; j < 10; ++j)
            {
                renderTexture[j].clear(sf::Color(0, 0, 0));

                for (int j = 0; j < 10; ++j)
                    renderTexture[j].draw(text);

                renderTexture[j].display();
            }

            window.clear(sf::Color(0, 0, 0));

            for (int j = 0; j < 10; ++j)
            {
                sprite.setTexture(renderTexture[j].getTexture());

                for (int k = 0; k < 10; ++k)
                    window.draw(sprite);
            }
        }

        window.display();

        sstr.str("");
        sstr << "Test -- Frame: " << clock.restart().asSeconds() << " sec";

        window.setTitle(sstr.str());
    }

    return EXIT_SUCCESS;
}
```
Frame times on my laptop dropped from 800ms down to 200ms with this change. This is a 4x speedup.
Everybody is free to test this change for bugs and performance deltas using real-world applications. I'd like feedback regarding more realistic workloads as well.