@roughbits01 2018-03-01
Tested it with my game which makes heavy use of RenderTextures (lightmap, glowing effect which needs multiple blur passes, heatmap, and other post effects) with a least 8 RenderTextures switches. Here is the result:

|                     |    Before      |    After    |
| ------------- |:-------------:| ----------:|
| CPU             | 12%             | 5%          |
| FPS              | 44                | 144         |