@binary1248 2018-03-01
@LaurentGomila Hiding the fact that FBOs are unshareable in RenderTextures that you can throw around at will is nothing short of black magic. 😁

I admit, this ended up becoming a bit of a complex optimization, but still... even if one doesn't understand the code at first glance, a cursory review should still possible (to catch typos, review documentation, etc.).