@myroidtc 2018-03-02
Tried it out. I can confirm it's a lot faster (don't have exact numbers) but there's some serious squish going on (comparison images below). Nothing changed except the SFML version. The other scenes in the game render fine, but they draw directly to the view. The simulation on the other hand is a RenderTexture's completed texture being drawn with a Sprite. The simulation's RenderTexture is 1280x720 but some 4096x4096 RenderTexture textures are drawn to it. The simulation's RenderTexture is also passed to a Shader Manager which applies bloom and other effects. It uses another couple of 1280x720s to do this.

![zeran s folly 2018-03-01 21-13-00-21](https://user-images.githubusercontent.com/5624103/36883379-1b132626-1d97-11e8-9c05-1ce46de66125.png)
![zeran s folly 2018-03-01 21-19-33-88](https://user-images.githubusercontent.com/5624103/36883380-1b2bebb6-1d97-11e8-9129-76a2757e7d81.png)
![zeran s folly 2018-03-01 21-14-26-19](https://user-images.githubusercontent.com/5624103/36883385-1e32408a-1d97-11e8-8009-a5d590f71f2f.png)
![zeran s folly 2018-03-01 21-21-54-49](https://user-images.githubusercontent.com/5624103/36883413-58cf494a-1d97-11e8-965a-dc6d4290e661.png)