@anthnich 2018-03-02
Can confirm that this is much, much faster. Ran 60 sec profiles for a before and after. The seconds noted are how much time that the function took to run during the 60s of execution:

RenderTexture Function|Before|After
--------------------------|-------|------
RenderWindow:Render|1.56s|0.23s
RenderWindow:RenderToFinal|3.30s|0.31s

**RenderWindow:Render**: Handles drawing of the Rendertextures that are utilized.
**RenderWindow:RenderToFinal**: Handles the final pass with any shader effects that need to be applied.

Bugs
---------

**Missing Render Texture**

I have a separate render texture that I used to draw the HUD and menu text in game. Then, it is last in the drawing order so it is drawn on top of everything. For some reason, the render texture is not drawing.

The overlay render texture uses sf::View and the size from another render texture to set it's size and position.

**Black Render Textures**

When I use my screenshot function (where I create a texture, update it from the renderwindow, copy it to the image, then save a file), a number of my render textures are black as well. These render textures that are black in the screenshot are showing up in game, but not the screenshot.

Also, the missing overlay "HUD" render texture mentioned above shows up the screenshot.