@binary1248 2018-03-03
@myroidtc @anthnich It would be helpful if you can provide a trace of a broken frame using something like [apitrace](http://apitrace.github.io/). Or... if you have the time, come up with a minimal code example that reproduces the problem, similar in form to the test case I provided above.

Also, if you haven't already done so, make sure to give the applications a go with debugging enabled and check to see if there are any OpenGL error messages output to the console.