@naezith 2018-03-03
I tested in couple different levels. It boosted FPS a lot!

500 -> 750
700 -> 1000
1000 -> 1150
1000 -> 1300

There is one issue quickly appeared, my water shader which reflects stuff, does not work. Waters are invisible. No shader errors in console. 

I draw them like this

```
// This is called once
          rtFBuffer.create(rt.getSize().x, rt.getSize().y);
          reflection = std::make_unique<sf::Sprite>(rtFBuffer.getTexture());

// After this line, these run every frame

        // Copy RT to Buffer to be the reflection
        {
            rt.display();

            sprFBuffer->setScale(1.0f, 1.0f);
            sprFBuffer->setPosition(0, 0);

            rtFBuffer.draw(*Game::i().sprFBuffer);

            rtFBuffer.display();
        }

        rt.resetGLStates();
        sf::Shader::bind(&shader_water);

        // Crop reflection areas and draw waters to main RT one by one
        for(auto& idx : watersInScreen) {
              reflection->setTextureRect(boundsOfWater);
              rt.draw(reflection);
        }

        sf::Shader::bind(nullptr);
```