@binary1248 2018-03-05
@naezith What is the performance difference with and without
```cpp
rt.resetGLStates();
sf::Shader::bind(&shader_water);
```
?

You are basically relying on undefined behaviour. SFML never guaranteed that it would not touch the shader binding within a call to `.draw()`. If it has to, it will, and you shouldn't rely on it staying this way in the future, independent from the changes in this pull request. The only guarantee that is made is that if you don't specify a shader in the `RenderStates` passed to `.draw()` that SFML with either disable a currently active shader or leave it deactivated if it is determined that none is active.