@naezith 2018-03-05
@binary1248 I didn't try that but giving shader to draw func as an argument was a lot slower, that's what I remember. Yet this works on Windows, Mac, Linux in every distro. Still I might try again.