@anthnich 2018-03-06
@binary1248 I sent you a PM on the SFML forums with the apitrace info.