@binary1248 2018-03-10
@anthnich @myroidtc @naezith Give the current version a try. I fixed several bugs.