@myroidtc 2018-03-10
Still getting the aspect ratio bug. In this scene, the background's RenderTexture (with the big circle) is drawn to the simulation's RenderTexture which is then drawn to the screen. The background's aspect ratio is even more exaggerated than the simulation's which gives me the impression that it's being applied to both of them.

![mana_normal](https://user-images.githubusercontent.com/5624103/37240618-49ed87dc-240b-11e8-8719-7243368a79e7.png)
![mana_stretch](https://user-images.githubusercontent.com/5624103/37240619-4a086e3a-240b-11e8-88b3-ff4be48a2dbd.png)