@MarioLiebisch 2018-03-10
Out of curiosity, what happens if you use a texture/screen resolution that's a power of two? like 1024x1024? Have you tried figuring out by which factor the whole thing is deformed?