@binary1248 2018-03-10
@myroidtc Can you provide an apitrace trace like @anthnich did or some kind of pseudo-code/snippet that demonstrates how you are drawing to the different RenderTargets? Working off of your current description, it is kind of hard to come up with a theory on why it is exhibiting this behaviour.