@binary1248 2018-03-10
@naezith The commit timestamp doesn't change if I amend it.