@naezith 2018-03-10
@binary1248 Okay I just tested it again. My shader issue is solved and pretty good FPS boost is still there but a bit less compared to old one I think, I'm not completely sure. I'll check if I have issues on Linux and Mac, if not, this will be in the next patch of my game. Thanks a lot!