@binary1248 2018-03-11
@myroidtc @anthnich @naezith Can you test the newest version? I redid the internal RenderTarget-to-Context tracking. It should take care of the edge cases as well now.

@LaurentGomila Do you think it is worth it to move GL state tracking out of RenderTarget and into e.g. a specialized class? It becomes really messy to cache states within the RenderTarget class when you start to have multiple RenderTargets all trying to live within the same context.